INTERNAL [IOTG] drm/i915: Introduce the Plane Restriction feature
This feature allows restrictions to the way planes are created in the
driver. Currently driver creates a primary plane and all the sprite
planes that are possible depending on the underlying hardware.This feature
allows restrictions on the number and type of planes that are created.
To accomodate the legacy APIs the first plane that is created is
designated as DRM_PLANE_TYPE_PRIMARY.
Usecase is to restrict the planes that are available for each
domain. Each domain passes in the plane mask through command line
paramenter- available_planes_per_pipe. Only these planes are created
for each domain. Another parameter domain_plane_owners is passed
into dom0 so that gvt is aware of the plane ownership for all the domains
and can restrict access.
DDBs are written once for all planes in Dom0 when in a virtualized
environment.

v2(ssingh) : Ported this patch to 4.19 kernel.
v3(Yakui)  : rebase it to 4.20 kernel and limit the plane restriction on
gvt_active or vgpu_active.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Signed-off-by: Anitha Chrisanthus <anitha.chrisanthus@intel.com>
Signed-off-by: Vivek Kasireddy <vivek.kasireddy@intel.com>
Change-Id: I293c490be8994bb525e511c8d1b8324b3c70932c